login/logout fix

+++ client/src/components/ProfileView/ProfileView.js
@@ -1,9 +1,10 @@
 import React, { useState, useEffect } from 'react'
+import { useDispatch } from 'react-redux';
 import { useLocation, useNavigate } from 'react-router-dom';
-import { useSelector } from 'react-redux';
 import SplitPane from 'react-split-pane';
 import axios from 'axios';
 
+import { logout } from '../../state/actions/userActions';
 import Recipe from '../Recipe/Recipe'
 import { API_BASE_URL, PATHS } from '../../Constants';
 import './style.css'
@@ -11,6 +12,7 @@ import './style.css'
 const ProfileView = ({ setUserToEdit }) => {
   const location = useLocation();
   const navigate = useNavigate();
+  const dispatch = useDispatch();
   const locationParts = location.pathname.split('/');
   const userId = locationParts[locationParts.length - 1];
   const [selectedUser, setSelectedUser] = useState(null);
@@ -47,10 +49,21 @@ const ProfileView = ({ setUserToEdit }) => {
 
   const deleteSelectedUser = (event) => {
     event.stopPropagation()
-    axios.delete(`${API_BASE_URL}/api/users/${userId}`)
-      .then(navigate(PATHS.HOME)
-      )
-      .catch(error => console.log('User Delete not successful', error))
+    const userAnswer = window.confirm(`If you delete your profile you will lost all your own recipes and all favorites recipes.
+    You won't be able to recover it! 
+    Are you absolutely sure you want to delete it?`)
+
+    if (userAnswer === true) {
+      axios.delete(`${API_BASE_URL}/api/users/${userId}`)
+        .then(navigate(PATHS.HOME)
+        )
+        .catch(error => console.log('User Delete not successful', error))
+    }
+
+    dispatch(logout());
+    window.localStorage.removeItem('user')
+    const authEvent = new Event('localStorageAuthEvent')
+    window.dispatchEvent(authEvent)
   }
 
   const onEditUser = (event) => {
